Refactor | Changed location of main_test.go, added scenario for stats_test.go and moved helper functions to helpers.go
