+++ tests/worker/main_integration/main_test.go
@@ -1,4 +1,4 @@
-package main
+package main_integration
 
 import (
 	"context"
@@ -22,6 +22,8 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
+
+//I am aware this is not the way we do integration tests now but this works, so I won't change it. FROM NOW ON WE DO IT RIGHT THO
 
 const (
 	DURATION       = 2 * time.Second
